refactor: logger middleware, all-exception-filter, main.ts
