+++ src/app.module.ts
@@ -1,5 +1,4 @@
 import { Module, MiddlewareConsumer } from '@nestjs/common';
-import { DataSource } from 'typeorm';
 import { UserModule } from './user/user.module';
 import { ArtistsModule } from './artists/artists.module';
 import { AlbumsModule } from './albums/albums.module';
@@ -12,9 +11,9 @@ import { AppService } from './app.service';
 import { AuthModule } from './auth/auth.module';
 import { APP_FILTER, APP_GUARD } from '@nestjs/core';
 import { JwtAuthGuard } from './auth/jwt-auth.guard';
-import { LoggerMiddleware } from './logger/logger.middleware';
+import { LoggerMiddleware } from './middleware/logger.middleware';
 import { CustomLoggerModule } from './logger/logger.module';
-import { AllExceptionFilter } from './logger/allExceptionsFilter';
+import { AllExceptionFilter } from './exception-filters/all-exception.filter';
 import { DatabaseModule } from './database/database.module';
 
 @Module({
@@ -40,7 +39,6 @@ import { DatabaseModule } from './database/database.module';
   ],
 })
 export class AppModule {
-  constructor(private dataSource: DataSource) {}
   configure(consumer: MiddlewareConsumer) {
     consumer.apply(LoggerMiddleware).forRoutes('*');
   }
