+++ src/logger/interfaces/log-data.interface.ts
